blog_dj/requirements.txt
asgiref==3.3.1
astroid==2.2.5
backcall==0.2.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.1
comtypes==1.1.7
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-bootstrap-form==3.4
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
docutils==0.16
googledrivedownloader==0.4
gunicorn==20.0.4
heroku==0.1.4
idm==0.38
idna==2.9
imutils==0.5.3
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jupyter-client==6.1.6
jupyter-core==4.6.3
Kivy==1.11.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
mahotas==1.4.11
matplotlib==3.3.0
mccabe==0.6.1
numpy==1.18.5
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
psycopg2==2.8.6
Pygments==2.6.1
pylint==2.3.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.0
pypiwin32==223
python-dateutil==1.5
python-decouple==3.3
pytz==2020.1
pywin32==228
pyzmq==19.0.1
reportlab==3.5.42
requests==2.23.0
scipy==1.5.4
selenium==3.141.0
six==1.12.0
soupsieve==2.0
sqlparse==0.4.1
static3==0.7.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.11.2
